% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_conver.R
\encoding{UTF-8}
\name{sem_conver}
\alias{sem_conver}
\title{SEM Direct, Indirect and Total Effects}
\usage{
sem_conver(sem=mod)
}
\arguments{
\item{sem}{a psem object}
}
\value{
an adjacency matrix
}
\description{
Converts a psem objet to a binary adjacency matrix
}
\examples{
\donttest{
## Not run:
#mod <- psem(
lm(rich ~ cover, data = keeley),
lm(cover ~ firesev, data = keeley),
lm(firesev ~ age, data = keeley),
data = keeley)

sem_conver(sem=mod)


##End
}
}
\author{
Anderson Medina
}
\keyword{sem}
