% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_ditter.R
\encoding{UTF-8}
\name{sem_ditter}
\alias{sem_ditter}
\title{SEM Direct, Indirect and Total Effects}
\usage{
sem_dittter(sem, response=NULL)
}
\arguments{
\item{sem}{a psem object}

\item{response}{a character vector with the name of the response variable}
}
\value{
a data.frame object in which rows are variables and columns are
direct, indirect and total effects.
}
\description{
Calculates direct, indirect and total effects from a piecewise SEM object.
Direct effects are the standardized coefficients, indirect effects are
calculated by multiplying the standardized coefficients of significant paths,
and total effects are the sum of the direct and indirect effects.
}
\examples{
\donttest{
## Not run:
mod <- psem(
lm(rich ~ cover, data = keeley),
lm(cover ~ firesev, data = keeley),
lm(firesev ~ age, data = keeley),
data = keeley)

sem_dittter(sem=mod, response="rich")

#Forgetting to set response variable:
sem_dittter(sem=mod, response=NULL)

##End
}
}
\author{
Anderson Medina
}
\keyword{sem}
