% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_guesser.R
\encoding{UTF-8}
\name{sem_guesser}
\alias{sem_guesser}
\title{SEM response variable guesser}
\usage{
sem_guesser(sem)
}
\arguments{
\item{sem}{a psem object}
}
\value{
a character vector with the name of the response variable
}
\description{
Tries to find the response variable in a psem object. Uses network approach
to measure variables degrees (in and out). The response is selected using two
criteria:
1) variables with smaller degrees out, in another words, variables that
predicts fewer or no variables. And if there are ties:
2) variables with higher degrees in, in another words, variables that are
explained by more or all variables.
}
\examples{
\donttest{
## Not run:
mod <- psem(
lm(rich ~ cover, data = keeley),
lm(cover ~ firesev, data = keeley),
lm(firesev ~ age, data = keeley),
data = keeley)

sem_guesser(sem=mod)

##End
}
}
\author{
Anderson Medina
}
\keyword{sem}
