% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_layouter.R
\encoding{UTF-8}
\name{sem_layouter}
\alias{sem_layouter}
\title{SEM Layout maker}
\usage{
sem_layouter(sem, response="NULL")
}
\arguments{
\item{sem}{a psem object}

\item{response}{a character vector with the name of the response variable}
}
\value{
a matrix with xy coordinates in the columns and variables in the rows
}
\description{
Tries to find a layout matrix using a user provided "response" variable, in
another words, the most endogenous variable in the structural equation model.
Puts the "response" variable on the right and places the other variables from
right (second most endogenous) to the left (exogenous or least endogenous).
}
\examples{
\donttest{
## Not run:
mod <- psem(
lm(rich ~ cover, data = keeley),
lm(cover ~ firesev, data = keeley),
lm(firesev ~ age, data = keeley),
data = keeley)

sem_layouter(sem=mod, response="rich")

##End
}
}
\author{
Anderson Medina
}
\keyword{sem}
