% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_plotter.R
\encoding{UTF-8}
\name{sem_plotter}
\alias{sem_plotter}
\title{SEM interactable Plot}
\usage{
sem_plotter(sem, stand, layout, ...)
}
\arguments{
\item{sem}{a psem object}

\item{stand}{Logical if TRUE uses standardized betas if FALSE uses raw betas}

\item{layout}{a coordinates matrix where each row is a variable}

\item{option}{a character vector of the type of graph. "full" plots paths
and correlated errors, "pred" omits correlated errors and "corr" plots only
the correlated errors}

\item{vertex.label}{a character vector to replace the variables names in the
plot. DEFAULT to NULL}

\item{vertex.color}{a character vector to define variable vertices colors}

\item{p.color}{character vector of length two with the colors of positive and
negative effects}

\item{c.color}{character vector of length one with the color of correlated
errors}
}
\value{
a list with two dataframes (edges and vertices), an \code{igraph}
object and a \code{visNetwork} object
}
\description{
Plots a structural equation model using a psem object created with the
\code{piecewise} package.
}
\examples{
\donttest{
## Not run:
mod <- psem(
lm(rich ~ cover, data = keeley),
lm(cover ~ firesev, data = keeley),
lm(firesev ~ age, data = keeley),
data = keeley)

lay<-matrix(data=c(3,2,1,4,1,2,1,2), nrow=4, ncol=2)

sem_plotter(sem=mod, layout=lay)

##End
}
}
\author{
Anderson Medina
}
\keyword{sem}
