% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_senser.R
\encoding{UTF-8}
\name{sem_senser}
\alias{sem_senser}
\title{SEM Direct, Indirect and Total Effects}
\usage{
sem_senser(x=NULL, dat=NULL, var.name=NULL, group=NULL, layout=NULL)
}
\arguments{
\item{x}{a square matrix containing the variables relationships}

\item{dat}{a data.frame containing the measured variables}

\item{sem}{a psem object}

\item{var.name}{group name as a character vector with length 1}

\item{layout}{a coordinates matrix where each row is a variable}

\item{method}{a character to define the method used. method="split" manually
separates the dataset into the number of groups and runs a psem for each
group. method=multigroup" calls multigroup from \code{piecewiseSEM} that
includes an interaction term in each regression}
}
\value{
a list with 1) the fitted psem models 2) \code{visNetwork} objects
}
\description{
Visual inspection of the sensibility of the SEM to different groups found
on the dataset.
}
\examples{
\donttest{
## Not run:
fac<-c(rep("A", 45), rep("B", 45))
keeley$group<-as.factor(fac)
sem <- psem(lm(rich ~ cover, data = keeley),
            lm(cover ~ firesev, data = keeley),
            lm(firesev ~ age, data = keeley),
data = keeley)

mat1 <- matrix(ncol=2, nrow=4)
mat1[1,1] <-3
mat1[1,2] <-1
mat1[2,1] <-2
mat1[2,2] <-2
mat1[3,1] <-1
mat1[3,2] <-1
mat1[4,1] <-4
mat1[4,2] <-2

# 1 Split with psem objet
sem_senser(sem=sem, var.name="group", method="split", layout=mat1)

# 2 Split with raw data

adj<-matrix(data=c(0,0,0,1,1,0,0,0,0,1,0,0,0,0,0,0), nrow=4, ncol=4, byrow = T)
colnames(adj)<-c("cover", "firesev", "age", "rich")
rownames(adj)<-c("cover", "firesev", "age", "rich")

sem_senser(x=adj, dat=keeley, var.name="group", method="split", layout=mat1)

# 3 Multigroup
sem_senser(sem=sem, var.name="group", method="multigroup", layout=mat1)

##End
}
}
\author{
Anderson Medina
}
\keyword{sem}
