% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_thinker.R
\encoding{UTF-8}
\name{sem_thinker}
\alias{sem_thinker}
\title{Creates a psem object based on an adjacency matrix}
\usage{
sem_thinker(x, dat)
}
\arguments{
\item{x}{a square matrix containing the variables relationships}

\item{dat}{a data.frame containing the measured variables}
}
\value{
a piecewise object
}
\description{
This functions creates a piecewise SEM based on two objects: a matrix
describing how variables are related to each other and a data.frame
containing the observed variables data.
}
\examples{
\donttest{
## Not run:

#Create or load a matrix
vals<-c(0,0,0,1,1,0,0,0,0,1,0,0,0,0,0,0)
mat<-matrix(data=vals, nrow=4, ncol=4, byrow = T)
colnames(mat)<-c("cover", "firesev", "age", "rich")
rownames(mat)<-c("cover", "firesev", "age", "rich")

#Creates the psem object based on the matrix
sem_thinker(x=mat, dat=keeley)

##End
}
}
\author{
Anderson Medina
}
\keyword{sem}
