% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_checker.R
\encoding{UTF-8}
\name{sem_checker}
\alias{sem_checker}
\title{Checks the assumptions of each linear equation of a SEM}
\usage{
sem_checker(x, dat, spatial=FALSE, coord=c("lon", "lat"))
}
\arguments{
\item{x}{a square matrix containing the variables relationships}

\item{dat}{a data.frame containing the measured variables}

\item{spatial}{logical evaluate spatial autocorrelation?}

\item{coord}{a character with geographic coordinates column names}
}
\value{
a list with original dataset and assumptions tests
}
\description{
This functions extracts the residuals of each linear equation and evaluates
the assumptions of homoscedasticity, normality, spatial autocorrelation and
multicolinearity
}
\examples{
\donttest{
## Not run:

#Create or load a matrix
vals<-c(0,0,0,1,1,0,0,0,0,1,0,0,0,0,0,0)
mat<-matrix(data=vals, nrow=4, ncol=4, byrow = T)
colnames(mat)<-c("cover", "firesev", "age", "rich")
rownames(mat)<-c("cover", "firesev", "age", "rich")

#Creates the psem object based on the matrix
sem_checker(x=mat, dat=keeley, spatial=F)

##End
}
}
\author{
Anderson Medina
}
\keyword{sem}
